refactor(order-products): extract shared OrderProducts component and groupProductItemsByShipmentId
Both account/order-detail and guest-order-lookup/order had identical local
copies. Move canonical implementation into app/components/order-products
and import from both pages. Removes ~80 lines of duplication.
